Simple email parsing agent/requirements.txt
langchain_openai
langgraph
grandalf
python-dotenv

